refactor: constants 폴더안에 index.js 생성
