+++ src/constants/index.js
@@ -1,0 +1,2 @@
+export { LOTTO, MATCH_COUNT, PROFIT, RANK, RESULT_TABLE } from './lotto';
+export { MESSAGE } from './messages';
